sequence inference limit 1000

+++ src/configs_utils.py
@@ -3,5 +3,5 @@
 
 
 def reduction_with_limit(df, col, limit=1000):
-    limit = max(len(df)*0.8, limit)
+    limit = int(max(len(df)*0.8, limit))
     return df.loc[df[col].nlargest(limit).index]
